Dynamically load module_template file for each module
Implement structure to create custom template functions for device templates
