+++ Modules/device/device_menu.php
@@ -4,4 +4,4 @@
     bindtextdomain($domain, "Modules/device/locale");
     bind_textdomain_codeset($domain, 'UTF-8');
 
-    $menu_dropdown_config[] = array('name'=> dgettext($domain, "Device Setup"), 'icon'=>'icon-facetime-video', 'path'=>"device/view" , 'session'=>"write", 'order' => 45 );
+    $menu_dropdown_config[] = array('name'=> dgettext($domain, "Device Setup"), 'icon'=>'icon-home', 'path'=>"device/view" , 'session'=>"write", 'order' => 45 );
